requirements.txt
kafka-python==2.0.2
boto3==1.28.62
awscli-local==0.20  # For interacting with LocalStack